fix compilation on gcc 4.1.2

+++ src/test/testAIO.cpp
@@ -34,7 +34,7 @@ void testPosixFdInput()
 	
 	PFIS.clear();
 	PFIS.seekg(0,std::ios::end);
-	assert ( PFIS.tellg() == static_cast<int64_t>(A.size()) );
+	assert ( static_cast<int64_t>(PFIS.tellg()) == static_cast<int64_t>(A.size()) );
 	PFIS.clear();
 	PFIS.seekg(0,std::ios::beg);
 	uint64_t const inc = 127;
